% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounds.R
\name{bounds}
\alias{bounds}
\title{Parameter bounds for glmCensRd()}
\usage{
bounds(Z = NULL, distY = "normal", distX = "normal")
}
\arguments{
\item{Z}{(optional) name(s) of additional fully observed covariates. If none, \code{Z = NULL} (the default).}

\item{distY}{distribution assumed for \code{Y} given \code{X} and \code{Z}. Default is \code{"normal"}, but other options are \code{"binomial"}, \code{"log-normal"}, \code{"weibull"}, \code{"exponential"}, or \code{"poisson"}.}

\item{distX}{distribution assumed for \code{X} given \code{Z}. Default is \code{"normal"}, but other options are \code{"log-normal"}, \code{"gamma"}, \code{"weibull"}, \code{"exponential"}, or \code{"poisson"}.}
}
\value{
A list with the following elements:
\item{lower}{a vector of the lower bounds for all parameters.}
\item{upper}{a vector of the upper bounds for all parameters.}
}
\description{
Parameter bounds for glmCensRd()
}
