% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_pXgivZ.R, R/d_pXgivZ.R
\name{calc_pXgivZ}
\alias{calc_pXgivZ}
\title{Calculate P(X|Z)}
\usage{
calc_pXgivZ(x, z = NULL, distX, eta_params)

calc_pXgivZ(x, z = NULL, distX, eta_params)
}
\arguments{
\item{x}{Predictor values (scalar or vector).}

\item{z}{Covariate values (scalar, vector, or dataframe).}

\item{distX}{Distribution assumed for \code{x} given \code{z}.}

\item{eta_params}{Vector of model parameters.}
}
\value{
A scalar or numeric vector the same length as the data input

A scalar or numeric vector the same length as the data input
}
\description{
Calculate P(X|Z)

Calculate P(X|Z)
}
