% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_pYgivXandZ.R
\name{calc_pYgivXandZ}
\alias{calc_pYgivXandZ}
\title{Calculate P(Y|X,Z)}
\usage{
calc_pYgivXandZ(y, x, z = NULL, distY, beta_params)
}
\arguments{
\item{y}{Outcome values (scalar or vector).}

\item{x}{Predictor values (scalar or vector).}

\item{z}{Covariate values (scalar, vector, or dataframe).}

\item{distY}{Distribution assumed for \code{y} given \code{x} and \code{z}.}

\item{beta_params}{Vector of model parameters.}
}
\value{
A scalar or numeric vector the same length as the data input
}
\description{
Calculate P(Y|X,Z)
}
