% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/completeCensRd.R
\name{completeCensRd}
\alias{completeCensRd}
\title{Create complete dataset}
\usage{
completeCensRd(Y, X, W, D, Z = NULL, partX = 50, data)
}
\arguments{
\item{Y}{Name of column variable.}

\item{X}{Name of censored predictor variable.}

\item{W}{Name of observed (i.e., censored) version of \code{X}.}

\item{D}{Name of event indicator, defined to be = 1 if \code{X} was uncensored.}

\item{Z}{(Optional) name(s) of additional fully observed covariates. Default is \code{NULL}.}

\item{partX}{Size of partition of unobserved \code{X} for censored subjects. Default is \code{50}.}

\item{data}{A dataframe containing at least columns \code{Y}, \code{X}, \code{W}, \code{D}, and \code{Z}.}
}
\value{
A dataframe containing the same columns as \code{data} but with duplicate rows for each censored subject with different potential \code{X} values.
}
\description{
Create complete dataset
}
