% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_vals.R
\name{init_vals}
\alias{init_vals}
\title{Initial values for glmCensRd()}
\usage{
init_vals(Z = NULL, distY = "normal", distX = "normal")
}
\arguments{
\item{Z}{(optional) name(s) of additional fully observed covariates. If none, \code{Z = NULL} (the default).}

\item{distY}{distribution assumed for \code{Y} given \code{X} and \code{Z}. Default is \code{"normal"}, but other options are \code{"binomial"}, \code{"log-normal"}, \code{"weibull"}, \code{"exponential"}, or \code{"poisson"}.}

\item{distX}{distribution assumed for \code{X} given \code{Z}. Default is \code{"normal"}, but other options are \code{"log-normal"}, \code{"gamma"}, \code{"weibull"}, \code{"exponential"}, or \code{"poisson"}.}
}
\value{
A vector of parameters
}
\description{
Initial values for glmCensRd()
}
