% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrateCensRd.R
\name{integrateCensRd}
\alias{integrateCensRd}
\title{Approximate the integral over censored observations in the log-likelihood using quadrature (the trapezoidal rule)}
\usage{
integrateCensRd(X, complete_data_cens)
}
\arguments{
\item{X}{Name of censored predictor variable.}

\item{complete_data_cens}{A dataframe containing artificially complete data for the censored subjects (returned from \code{completeCensRd()})}
}
\value{
A vector containing the approximate integral over the joint density of censored subjects from \code{complete_data_cens}.
}
\description{
Approximate the integral over censored observations in the log-likelihood using quadrature (the trapezoidal rule)
}
