% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg_am.R
\name{linreg_am}
\alias{linreg_am}
\title{Maximum likelihood estimator (MLE) for right-censored covariates in normal linear regression}
\usage{
linreg_am(params0, Y, X, C, Z = NULL, data, steptol = 1e-06, iterlim = 100)
}
\arguments{
\item{params0}{Starting parameter values.}

\item{Y}{Name of column variable.}

\item{X}{Name of censored covariate variable.}

\item{C}{Name of observed version of \code{X}.}

\item{Z}{(Optional) name(s) of additional fully observed covariates. Default is \code{Z = NULL}}

\item{data}{A dataframe containing at least columns \code{Y}, \code{X}, \code{C}, \code{Z}.}

\item{steptol}{(Fed to \code{nlm()}) A positive scalar providing the minimum allowable relative step length. Default is \code{steptol = 1e-6}.}

\item{iterlim}{(Fed to \code{nlm()}) A positive integer specifying the maximum number of iterations to be performed before the program is terminated. Default is \code{iterlim = 100}.}
}
\value{
A list with the following two elements:
\item{coeff}{A dataframe containing coefficient and standard error estimates at convergence.}
\item{code}{An integer indicating why the optimization process terminated. See \code{?nlm} for details on values.}
}
\description{
Maximum likelihood estimator (MLE) for right-censored covariates in normal linear regression
}
