% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik.R
\name{loglik}
\alias{loglik}
\title{Observed-data log-likelihood}
\usage{
loglik(
  params,
  Y,
  X,
  W,
  D,
  Z = NULL,
  partX = 50,
  distY = "normal",
  distX = "normal",
  data
)
}
\arguments{
\item{params}{Parameter values.}

\item{Y}{Name of outcome variable.}

\item{X}{Name of censored predictor variable.}

\item{W}{Name of observed (i.e., censored) version of \code{X}.}

\item{D}{Name of event indicator, defined to be = 1 if \code{X} was uncensored.}

\item{Z}{(Optional) name(s) of additional fully observed covariates. Default is \code{NULL}}

\item{partX}{Size of partition of unobserved \code{X} for censored subjects. Default is \code{50}.}

\item{distY}{Distribution assumed for \code{Y} given \code{X} and \code{Z}. Default is \code{"normal"}.}

\item{distX}{Distribution assumed for \code{X} given \code{Z}. Default is \code{"normal"}.}

\item{data}{A dataframe containing at least columns \code{Y}, \code{X}, \code{W}, \code{D}, and \code{Z} (if applicable).}
}
\value{
A scalar of the log-likelihood function (negated for use with \code{nlm}()).
}
\description{
Observed-data log-likelihood
}
