% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_cc.R
\name{loglik_cc}
\alias{loglik_cc}
\title{Complete-case log-likelihood for GLMs}
\usage{
loglik_cc(params, Y, X, Z = NULL, data, distY = "normal", distX = "normal")
}
\arguments{
\item{params}{parameter values.}

\item{Y}{name of outcome variable.}

\item{X}{name of censored predictor variable.}

\item{Z}{(optional) name(s) of additional fully observed covariates. If none, \code{Z = NULL} (the default).}

\item{data}{a dataframe containing at least columns \code{Y}, \code{X}, \code{C}, \code{Z}.}

\item{distY}{distribution assumed for \code{Y} given \code{X} and \code{Z}. Default is \code{"normal"}, but other options are \code{"binomial"}, \code{"log-normal"}, \code{"weibull"}, \code{"exponential"}, or \code{"poisson"}.}

\item{distX}{distribution assumed for \code{X} given \code{Z}. Default is \code{"normal"}, but other options are \code{"log-normal"}, \code{"gamma"}, \code{"inverse-gaussian"}, \code{"weibull"}, \code{"exponential"}, or \code{"poisson"}.}
}
\value{
A scalar of the log-likelihood function (negated for use with \code{nlm}()).
}
\description{
Complete-case log-likelihood for GLMs
}
