% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_normal_normal.R
\name{loglik_normal_normal}
\alias{loglik_normal_normal}
\title{Observed-data log-likelihood for normal linear regression with a censored predictor modeled with normal linear regression, also.}
\usage{
loglik_normal_normal(params, Y, X, W, D, Z = NULL, partX = 50, data)
}
\arguments{
\item{Y}{Name of outcome variable.}

\item{X}{Name of censored predictor variable.}

\item{W}{Name of observed (i.e., censored) version of \code{X}.}

\item{D}{Name of event indicator, defined to be = 1 if \code{X} was uncensored.}

\item{Z}{(Optional) name(s) of additional fully observed covariates. Default is \code{NULL}}

\item{partX}{Size of partition of unobserved \code{X} for censored subjects. Default is \code{50}.}

\item{data}{A dataframe containing at least columns \code{Y}, \code{X}, \code{W}, \code{D}, and \code{Z}.}

\item{params0}{Parameter values.}
}
\value{
A vector containing the approximate integral over the joint density of censored subjects from \code{complete_data_cens}.
}
\description{
Observed-data log-likelihood for normal linear regression with a censored predictor modeled with normal linear regression, also.
}
