% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_deriv2_loglik.R
\name{calc_deriv2_loglik}
\alias{calc_deriv2_loglik}
\title{Calculate the second derivatives of the individual observations' log-likelihood contributions}
\usage{
calc_deriv2_loglik(
  params,
  Y,
  X,
  W,
  D,
  Z = NULL,
  subdivisions = 100,
  distY = "normal",
  distX = "normal",
  data
)
}
\arguments{
\item{params}{parameter values.}

\item{Y}{name of outcome variable.}

\item{X}{name of censored predictor variable.}

\item{W}{name of observed (censored) version of \code{X}.}

\item{D}{name of event indicator, defined to be \code{= 1} if \code{X} was uncensored and \code{0} otherwise.}

\item{Z}{(optional) name(s) of additional fully observed covariates. If none, \code{Z = NULL} (the default).}

\item{subdivisions}{(fed to \code{integrate}) the maximum number of subintervals used to integrate over unobserved \code{X} for censored subjects. Default is \code{100}.}

\item{distY}{distribution assumed for \code{Y} given \code{X} and \code{Z}. Default is \code{"normal"}, but \code{"binomial"} is the other option.}

\item{distX}{distribution assumed for \code{X} given \code{Z}. Default is \code{"normal"}, but other options are \code{"log-normal"}, \code{"gamma"}, \code{"inverse-gaussian"}, \code{"weibull"}, \code{"exponential"}, or \code{"poisson"}.}

\item{data}{a dataframe containing at least columns \code{Y}, \code{X}, \code{C}, \code{Z}.}
}
\value{
A vector containing the derivatives of the log-likelihood contributions of the rows/observations in \code{data}.
}
\description{
Calculate the second derivatives of the individual observations' log-likelihood contributions
}
