% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_pYXandZ.R
\name{calc_pYXandZ}
\alias{calc_pYXandZ}
\title{Calculate P(Y,X,Z) = P(Y|X,Z)P(X|Z)}
\usage{
calc_pYXandZ(x, y, z = NULL, lengthZ = 0, distY, distX, params)
}
\arguments{
\item{x}{predictor values (scalar or vector).}

\item{y}{outcome values (scalar or vector).}

\item{z}{covariate values (scalar, vector, or dataframe).}

\item{lengthZ}{number of additional covariates. If none (the default), \code{lengthZ = 0}.}

\item{distY}{distribution assumed for \code{y} given \code{x} and \code{z}.}

\item{distX}{distribution assumed for \code{x} given \code{z}.}

\item{params}{parameter values.}
}
\value{
A scalar or numeric vector the same length as the data input
}
\description{
Calculate P(Y,X,Z) = P(Y|X,Z)P(X|Z)
}
