% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmCensRd.R
\name{glmCensRd}
\alias{glmCensRd}
\title{Maximum likelihood estimator (MLE) for censored predictor in generalized linear models (GLM)}
\usage{
glmCensRd(
  Y,
  W,
  D,
  Z = NULL,
  partX = 50,
  distY = "normal",
  distX = "normal",
  data,
  steptol = 1e-04,
  iterlim = 100
)
}
\arguments{
\item{Y}{Name of column variable.}

\item{W}{Name of observed (i.e., censored) version of \code{X}.}

\item{D}{Name of event indicator, defined to be = 1 if \code{X} was uncensored.}

\item{Z}{(Optional) name(s) of additional fully observed covariates. Default is \code{NULL}}

\item{partX}{Size of partition of unobserved \code{X} for censored subjects. Default is \code{50}.}

\item{distY}{Distribution assumed for \code{Y} given \code{X} and \code{Z}. Default is \code{"normal"}.}

\item{distX}{Distribution assumed for \code{X} given \code{Z}. Default is \code{"normal"}.}

\item{data}{A dataframe containing at least columns \code{Y}, \code{X}, \code{C}, \code{Z}.}

\item{steptol}{(Fed to \code{nlm()}) A positive scalar providing the minimum allowable relative step length. Default is \code{1e-4}.}

\item{iterlim}{(Fed to \code{nlm()}) A positive integer specifying the maximum number of iterations to be performed before the program is terminated. Default is \code{100}.}
}
\value{
A list with the following elements:
\item{outcome_model}{A list containing details of the fitted model for the outcome.}
\item{predictor_model}{A list containing details of the fitted model for the predictor.}
\item{code}{An integer indicating why the optimization process terminated. See \code{?nlm} for details on values.}
}
\description{
Maximum likelihood estimator (MLE) for censored predictor in generalized linear models (GLM)
}
