% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmCensRd.R
\name{glmCensRd}
\alias{glmCensRd}
\title{Maximum likelihood estimator (MLE) for censored predictor in generalized linear models (GLM)}
\usage{
glmCensRd(
  Y,
  W,
  D,
  Z = NULL,
  data,
  distY = "normal",
  distX = "normal",
  cens = "right",
  robcov = TRUE,
  subdivisions = 100,
  steptol = 1e-06,
  iterlim = 100,
  verbose = FALSE
)
}
\arguments{
\item{Y}{name of outcome variable.}

\item{W}{name of observed (censored) version of \code{X}.}

\item{D}{name of event indicator, defined to be \code{= 1} if \code{X} was uncensored and \code{0} otherwise.}

\item{Z}{(optional) name(s) of additional fully observed covariates. If none, \code{Z = NULL} (the default).}

\item{data}{dataframe containing at least columns \code{Y}, \code{X}, \code{C}, \code{Z}.}

\item{distY}{distribution assumed for \code{Y} given \code{X} and \code{Z}. Default is \code{"normal"}, but other options are \code{"binomial"}, \code{"log-normal"}, \code{"weibull"}, \code{"exponential"}, or \code{"poisson"}.}

\item{distX}{distribution assumed for \code{X} given \code{Z}. Default is \code{"normal"}, but other options are \code{"log-normal"}, \code{"gamma"}, \code{"weibull"}, \code{"exponential"}, or \code{"poisson"}.}

\item{cens}{type of censoring assumed for \code{X}. Default is \code{"right"}, but the other option is \code{"left"}.}

\item{robcov}{logical. If \code{TRUE} (the default), the robust sandwich covariance estimator of parameter standard errors is included in the output.}

\item{subdivisions}{(fed to \code{integrate}) the maximum number of subintervals used to integrate over unobserved \code{X} for censored subjects. Default is \code{100}.}

\item{steptol}{(fed to \code{nlm}) a positive scalar providing the minimum allowable relative step length. Default is \code{1E-6}.}

\item{iterlim}{(fed to \code{nlm}) a positive integer specifying the maximum number of iterations to be performed before the program is terminated. Default is \code{100}.}

\item{verbose}{logical. If \code{TRUE}, text is printed to show progress. Default is \code{FALSE}.}
}
\value{
A list with the following elements:
\item{outcome_model}{a list containing details of the fitted model for the outcome.}
\item{predictor_model}{a list containing details of the fitted model for the predictor.}
\item{code}{an integer indicating why the optimization process terminated. See \code{?nlm} for details on values.}
\item{vcov}{variance-covariance matrix of the model parameters.}
\item{rvcov}{robust sandwich variance-covariance matrix of the model parameters.}
}
\description{
Maximum likelihood estimator (MLE) for censored predictor in generalized linear models (GLM)
}
