% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_normal_normal.R
\name{loglik_normal_normal}
\alias{loglik_normal_normal}
\title{Observed-data log-likelihood for normal linear regression with a censored predictor modeled with normal linear regression, also.}
\usage{
loglik_normal_normal(params, Y, X, Z = NULL, complete_data)
}
\arguments{
\item{Y}{Name of outcome variable.}

\item{X}{Name of censored predictor variable.}

\item{Z}{(Optional) name(s) of additional fully observed covariates. Default is \code{NULL}}

\item{params0}{Parameter values.}

\item{W}{Name of observed (i.e., censored) version of \code{X}.}

\item{data}{A dataframe containing at least columns \code{Y}, \code{X}, \code{W}, \code{Z}.}

\item{steptol}{(Fed to \code{nlm()}) A positive scalar providing the minimum allowable relative step length. Default is \code{steptol = 1e-6}.}

\item{iterlim}{(Fed to \code{nlm()}) A positive integer specifying the maximum number of iterations to be performed before the program is terminated. Default is \code{iterlim = 100}.}
}
\value{
A vector containing the approximate integral over the joint density of censored subjects from \code{complete_data_cens}.
}
\description{
Observed-data log-likelihood for normal linear regression with a censored predictor modeled with normal linear regression, also.
}
